Add fast LZO compressor
This is not a stream compressor: it just caches everything in the
stream, then calls LibLZO.compress on the buffer. Need to work out how
to deal with encoding first literals smaller than 4 bytes into a stream
of LZO chunks before this can be made into a sort-of streaming
interface.